Fix failing RPMS scenario $$LIST^ZISH test
It seems that due to stack handling improvements/fixes in YottaDB, we
now see that the stack was not completely unwound when we get into the
error condition when we $ZPARSE the directory, resulting in continuing
in LIST even after $$DEFDIR has failed. This is now fixed by poping the
stack properly. This fix mirrors the fix made in
01ba7229f9ec20ae619d2fa00d5ef4ee341848b3.