SAGG Data Collection was too slow.
It turns out that most of the time was being spent in the grep calls.
Now we refactored that code so that instead of grepping, we create an
array containing the number of sets and kills. This is much faster, as
we only have go through the file once rather than an estimated 800
times.