KMPLOG -> KMPVLOG, Del Old Logs, Proc Name
All references to KMPLOG d to KMPVLOG to better align with the VSM
namespace.

KMPVLOG now uses HL7 dates instead of Fileman dates

Per Bhaskar's suggestion, all old logs are now deleted per either
retention parameters in VSM or hardcoded to 1 day for cover sheet loads or 56
days for SAGG data. KMPVLOG has DELLOG tag, and a unit test is supplied.

cv-load file is now cvload, because I use the - as a delimiter in
finding out the date of the file.

Calls to SETNM^%ZOSV in VTCM as we don't know the name of the user so
that we get a process name in the ^ZSY listing.