Patch level checker update.
Option CP Environment Check [KMPD STATUS] checks patch level for
routines for the package. I update this so that the routines come out
checked out correctly and add support for it to get the correct patch
level for the percent routines, which can have different patches
depending on which M runtime on you are on.